% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_process_subfuncs_omop.R
\name{check_code_dist_csd_omop}
\alias{check_code_dist_csd_omop}
\title{Base CSD function}
\usage{
check_code_dist_csd_omop(
  cohort_codedist,
  concept_set,
  time = FALSE,
  time_span,
  time_period,
  domain_tbl = conceptsetdistribution::csd_domain_file
)
}
\arguments{
\item{cohort_codedist}{the cohort of interest}

\item{concept_set}{an annotated concept set CSV file with the following columns:
\itemize{
\item \code{concept_id} required for OMOP; the concept_id of interest
\item \code{concept_code} required for PCORnet; the code of interest
\item \code{concept_name} optional; the descriptive name of the concept
\item \code{vocabulary_id} required for PCORnet; the vocabulary of the code - should match what is listed in the domain table's vocabulary_field
\item \code{variable} required; a string label grouping one concept code into a larger variable definition
\item \code{domain} required; the name of the CDM table where the concept is stored - multiple domains can be included in the file, but only one domain should be listed per row
}}

\item{time}{logical to determine whether to output the check across time}

\item{time_span}{when time = TRUE, a vector of two dates for the observation period of the study}

\item{time_period}{when time = TRUE, this argument defines the distance between dates within the specified time period. defaults
to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}

\item{domain_tbl}{input table defining the domains listed in the annotated concept set
four columns:
- \code{domain} the name of the CDM table associated with the concept; should match what is listed in the annotated concept set
- \code{concept_field} the name of the field in the domain table where the concepts are located
- \code{date_field} the name of the field in the domain table with the date that should be used for time-based filtering
- \code{vocabulary_field} PCORnet only; set to NA}
}
\value{
returns variable and its concept mappings, both in counts and in proportions;
when time = TRUE, then output is given across time, and proportions computed within each variable
}
\description{
Base CSD function
}
\keyword{internal}
