% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_csd_pcnt.R
\name{check_code_dist_csd_pcnt}
\alias{check_code_dist_csd_pcnt}
\title{Base CSD function}
\usage{
check_code_dist_csd_pcnt(
  cohort_codedist,
  concept_set,
  time = FALSE,
  time_span,
  time_period,
  domain_tbl = read_codeset("scv_domains", "cccc")
)
}
\arguments{
\item{cohort_codedist}{the cohort to pass in}

\item{concept_set}{the concept set passed in through \code{csd_process};
concept set CSV file with the following columns:
\code{concept_id} | \code{concept_code} | \code{concept_name} | \code{vocabulary_id} | \code{category} | \code{variable} | \code{domain}
The variable field is required to categorize each concept set into a particular variable
The domain is required so that the function knows which table to join to in order to derive counts}

\item{time}{logical to determine whether to output the check across time}

\item{time_span}{when \code{time = TRUE}, a vector of two dates for the observation period of the study}

\item{time_period}{when time = TRUE, this argument defines the distance between dates within the specified time period. defaults
to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}

\item{domain_tbl}{domain table passed in through \code{csd_process};
tbl that is similar to the SCV check;
four columns: \code{domain} | \code{source_col} | \code{concept_col} | \code{date_col};
the required columns for the csd check are only \code{domain_tbl}, \code{concept_col}, \code{date_col}}
}
\value{
returns variable and their concept mappings, both in counts and in proportions;
when \code{time = TRUE}, then output is given across time, and proportions computed within each variable
}
\description{
Base CSD function
}
