% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_process_subfuncs.R
\name{check_code_dist_ssanom_omop}
\alias{check_code_dist_ssanom_omop}
\title{Base CSD function only for \verb{single site, anomaly, no time}}
\usage{
check_code_dist_ssanom_omop(
  cohort_codedist,
  concept_set,
  num_concept_combined = FALSE,
  num_concept_1 = 30,
  num_concept_2 = 30,
  domain_tbl = conceptsetdistribution::csd_domain_file
)
}
\arguments{
\item{cohort_codedist}{the cohort to pass in}

\item{concept_set}{the concept set passed in through \code{csd_process};
concept set CSV file with the following columns:
\code{concept_id} | \code{concept_code} | \code{concept_name} | \code{vocabulary_id} | \code{category} | \code{variable} | \code{domain}
The variable field is required to categorize each concept set into a particular variable
The domain is required so that the function knows which table to join to in order to derive counts}

\item{num_concept_combined}{when \code{mult_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument is an integer and will ensure that \code{concept1} and \code{concept2} meet
some minimal threshold for including in the jaccard index; if \code{TRUE}, then
\emph{both} conditions for \code{num_concept_1} and \code{num_concept_2} should be met;
if \code{FALSE} then just one condition needs to be met.}

\item{num_concept_1}{when \code{mult_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument is an integer and requires a minimum number of times that
the \emph{first} concept appears in the dataset}

\item{num_concept_2}{when \code{mult_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument is an integer and requires a minimum number of times that
the \emph{second} concept appears in the dataset}

\item{domain_tbl}{domain table passed in through \code{csd_process};
tbl that is similar to the SCV check;
four columns: \code{domain} | \code{source_col} | \code{concept_col} | \code{date_col};
the required columns for the csd check are only \code{domain_tbl}, \code{concept_col}, \code{date_col}}
}
\value{
the jaccard index of two different concepts for a given variable
}
\description{
Base CSD function only for \verb{single site, anomaly, no time}
}
