% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csd_concept_set}
\alias{csd_concept_set}
\title{CSD Sample Concept Set}
\format{
\subsection{\code{csd_concept_set}}{

A data frame with 6 columns
\describe{
\item{concept_id}{The OMOP concept_id; if the PCORnet CDM is being used, default this column to a random integer like the row number}
\item{concept_code}{The original code associated with the concept_id}
\item{concept_name}{(optional)The string name of the concept}
\item{vocabulary_id}{The vocabulary associated with the concept; if the PCORnet CDM is being used, ensure that the values of this field match the vocabulary abbreviations used in the CDM itself}
\item{variable}{A string label for the variable associated with the concept}
\item{domain}{The domain table where the concept should be identified. This should match the domain listed in the domain_tbl file}
}
}
}
\usage{
csd_concept_set
}
\description{
A sample version of the file structure expected for the \code{concept_set}
parameted in the \code{csd_process} function. The user should recreate this
file and include their own clinical concepts.
}
\keyword{datasets}
