% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csd_domain_file}
\alias{csd_domain_file}
\title{CSD Sample Domain File}
\format{
\subsection{\code{csd_domain_file}}{

A data frame with 3 rows and 3 columns
\describe{
\item{domain}{The name of the CDM table associated with the domain where the concept of interest can be found. Should match the domain listed in the concept_set file.}
\item{concept_field}{The name of the column in the domain table that contains the concepts of interest listed in the concept_set file.}
\item{date_field}{The name of the column in the domain table that contains dates to be used for time-based filtering.}
\item{vocabulary_field}{(PCORnet only) The name of the column in the domain table where the vocabulary type is stored}
}
}
}
\usage{
csd_domain_file
}
\description{
A sample version of the file structure expected for the \code{domain_tbl}
parameter in the \code{csd_process} function. The user should recreate
this file and include their own domain definitions.
}
\keyword{datasets}
