% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_csd_output.R
\name{csd_ms_anom_at}
\alias{csd_ms_anom_at}
\title{\strong{Multi-Site Across Time Anomaly}}
\usage{
csd_ms_anom_at(process_output, filter_concept, concept_col)
}
\arguments{
\item{process_output}{output from \code{csd_process}}

\item{filter_concept}{the concept_id that should be used for the output}
}
\value{
three graphs:
\enumerate{
\item Loess smoothed line graph that shows the proportion of a code across time
with the Euclidean Distance associated with each line
\item same as (1) but displaying the raw, unsmoothed proportion
\item a radial bar graph displaying the Euclidean Distance value for each
site, where the color is the average proportion across time
}

THIS GRAPH SHOWS ONLY ONE CONCEPT AT A TIME!
}
\description{
\strong{Multi-Site Across Time Anomaly}
}
