% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_output_subfuncs.R
\name{csd_ms_anom_cs}
\alias{csd_ms_anom_cs}
\title{\emph{Multi-Site, Anomaly, Cross-Sectional}}
\usage{
csd_ms_anom_cs(
  process_output,
  concept_col,
  text_wrapping_char = 80,
  filtered_var,
  comparison_col = "prop_concept",
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{output from \code{csd_process}}

\item{concept_col}{the name of the column from the concept_set used to identify concepts
should be either \code{concept_id} or \code{concept_code}}

\item{text_wrapping_char}{the number of characters for the \code{concept_name} or \code{concept_id} to
display on heatmap; limited to 80}

\item{filtered_var}{the variable to perform the analysis on from the data frame; column name
that contains the variable names should be labeled \code{variable}}

\item{comparison_col}{the column that computes the quantitative value for comparison across sites;
in \code{csd} check, it is the \code{prop_concept}}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the proportion of usage for
a given code, and the size of the dot represents the mean code
usage across all sites
}
\description{
\emph{Multi-Site, Anomaly, Cross-Sectional}
}
