% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_output_subfuncs.R
\name{csd_ms_anom_la}
\alias{csd_ms_anom_la}
\title{\strong{Multi-Site, Anomaly, Longitudinal}}
\usage{
csd_ms_anom_la(
  process_output,
  filter_concept,
  concept_col,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{output from \code{csd_process}}

\item{filter_concept}{the concept_id that should be used for the output}

\item{concept_col}{the name of the column from the concept_set used to identify concepts
should be either \code{concept_id} or \code{concept_code}}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
three graphs:
\enumerate{
\item Loess smoothed line graph that shows the proportion of a code across time
with the Euclidean Distance associated with each line
\item same as (1) but displaying the raw, unsmoothed proportion
\item a radial bar graph displaying the Euclidean Distance value for each
site, where the color is the average proportion across time
}

THIS GRAPH SHOWS ONLY ONE CONCEPT AT A TIME!
}
\description{
\strong{Multi-Site, Anomaly, Longitudinal}
}
