% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_csd_output.R
\name{csd_ms_exp_nt}
\alias{csd_ms_exp_nt}
\title{\emph{Multi Site, Exploratory, No Time}}
\usage{
csd_ms_exp_nt(process_output, facet, concept_col, num_codes = 10)
}
\arguments{
\item{process_output}{dataframe output by \code{csd_process}}

\item{facet}{the variables by which you would like to facet the graph;
defaults to NULL}

\item{num_codes}{the number of top codes per variable that should be
displayed in the table}
}
\value{
a searchable and filterable table with mappings, proportion of representation, and
denominator counts for the number of codes selected
in @num_codes
}
\description{
\emph{Multi Site, Exploratory, No Time}
}
