% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_csd_output.R
\name{csd_ss_anom_at}
\alias{csd_ss_anom_at}
\title{\emph{Single Site, Anomaly, Across Time}}
\usage{
csd_ss_anom_at(
  process_output,
  concept_col,
  filtered_var = "ibd",
  filter_concept = 81893,
  facet = NULL
)
}
\arguments{
\item{process_output}{dataframe output by \code{csd_process}}

\item{filtered_var}{the variable to perform the anomaly detection for}

\item{filter_concept}{the concept_id of interest for the analysis}

\item{facet}{the variables by which you would like to facet the graph; defaults to NULL}
}
\value{
if analysis was executed by year or greater, a P Prime control chart
is returned with outliers marked with orange dots

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if analysis was executed by month or smaller, an STL regression is 
    conducted and outliers are marked with red dots. the graphs representing
    the data removed in the regression are also returned
}\if{html}{\out{</div>}}
}
\description{
Control chart looking at number of mappings over time
}
