% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_output_subfuncs.R
\name{csd_ss_anom_cs}
\alias{csd_ss_anom_cs}
\title{\emph{Single Site, Anomaly, Cross-Sectional}}
\usage{
csd_ss_anom_cs(process_output, vocab_tbl = NULL, filtered_var)
}
\arguments{
\item{process_output}{the output from \code{csd_process}}

\item{vocab_tbl}{OPTIONAL: the location of an external vocabulary table containing concept names for
the provided codes. if not NULL, concept names will be available in either a reference
table or in a hover tooltip}

\item{filtered_var}{the variable to perform the jaccard similarity index for}
}
\value{
for a given variable, a heatmap of the jaccard index for each concept pair
}
\description{
\emph{Single Site, Anomaly, Cross-Sectional}
}
