% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_output_subfuncs.R
\name{csd_ss_anom_la}
\alias{csd_ss_anom_la}
\title{\emph{Single Site, Anomaly, Longitudinal}}
\usage{
csd_ss_anom_la(
  process_output,
  concept_col,
  filtered_var,
  filter_concept,
  facet = NULL
)
}
\arguments{
\item{process_output}{dataframe output by \code{csd_process}}

\item{concept_col}{the name of the column from the concept_set used to identify concepts
should be either \code{concept_id} or \code{concept_code}}

\item{filtered_var}{the variable to perform the anomaly detection for}

\item{filter_concept}{the concept_id of interest for the analysis}

\item{facet}{the variables by which you would like to facet the graph; defaults to NULL}
}
\value{
if analysis was executed by year or greater, a P Prime control chart
is returned with outliers marked with orange dots

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if analysis was executed by month or smaller, an STL regression is
    conducted and outliers are marked with red dots. the graphs representing
    the data removed in the regression are also returned
}\if{html}{\out{</div>}}
}
\description{
\emph{Single Site, Anomaly, Longitudinal}
}
\keyword{internal}
