% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_csd_output.R
\name{csd_ss_exp_at}
\alias{csd_ss_exp_at}
\title{\emph{Single Site, Exploratory, Across Time}}
\usage{
csd_ss_exp_at(
  process_output,
  concept_col,
  facet = NULL,
  filtered_var = c("spondyloarthritis"),
  num_mappings = 10,
  output_value = "prop_concept"
)
}
\arguments{
\item{process_output}{dataframe output by \code{csd_process}}

\item{facet}{the variables by which you would like to facet the graph;
defaults to NULL}

\item{filtered_var}{the variable to perform the anomaly detection for}

\item{num_mappings}{an integer indicating the number of top codes for the
filtered_var of interest that should be displayed}

\item{output_value}{the numerical column in the data that should be displayed}
}
\value{
a line graph with one facet per variable displaying the proportion of mapped codes
across the user selected time period

a reference table with total counts of each code across the entire user selected
time period
}
\description{
\emph{Single Site, Exploratory, Across Time}
}
