% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_output_subfuncs.R
\name{csd_ss_exp_cs}
\alias{csd_ss_exp_cs}
\title{\emph{Single Site, Exploratory, Cross-Sectional}}
\usage{
csd_ss_exp_cs(
  process_output,
  concept_col,
  facet = NULL,
  num_codes = 10,
  num_mappings = 10
)
}
\arguments{
\item{process_output}{the output from \code{csd_process}}

\item{concept_col}{the name of the column from the concept_set used to identify concepts
should be either \code{concept_id} or \code{concept_code}}

\item{facet}{variables to facet by; defaults to NULL}

\item{num_codes}{an integer to represent the top number of codes to include in the mappings for the exploratory analyses;
will pick the codes based on the highest count of the most commonly appearing variables;}

\item{num_mappings}{an integer to represent the top number of mappings for a given variable in the exploratory analyses}
}
\value{
a list with two elements:
1) heatmap for to \code{n} concepts (\code{num_codes})  and \code{x} variables (\code{num_mappings}),
with proportion for each concept.
2) a table with each mapping and the total variable count
}
\description{
\emph{Single Site, Exploratory, Cross-Sectional}
}
\keyword{internal}
