% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_csd_output.R
\name{csd_ms_anom_nt}
\alias{csd_ms_anom_nt}
\title{\emph{Multi-Site Anomaly No Time}}
\usage{
csd_ms_anom_nt(
  process_output,
  concept_col,
  text_wrapping_char = 80,
  filtered_var = "ibd",
  comparison_col = "prop_concept"
)
}
\arguments{
\item{process_output}{output from \code{csd_process}}

\item{text_wrapping_char}{the number of characters for the \code{concept_name} or \code{concept_id} to
display on heatmap; limited to 80}

\item{filtered_var}{the variable to perform the analysis on from the data frame; column name
that contains the variable names should be labeled \code{variable}}

\item{comparison_col}{the column that computes the quantitative value for comparison across sites;
in \code{csd} check, it is the \code{prop_concept}}

\item{grouped_vars}{a vector containing the variables to group by to compute the mean across the sites;
in \code{csd} check, defaulted to c(\code{variable}, \code{concept_id})}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the proportion of usage for
a given code, and the size of the dot represents the mean code
usage across all sites
}
\description{
\emph{Multi-Site Anomaly No Time}
}
