% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_output.R
\name{csd_output}
\alias{csd_output}
\title{Concept Set Distribution -- Output Generation}
\usage{
csd_output(
  process_output,
  output_function,
  concept_set = NULL,
  vocab_tbl = NULL,
  num_variables = 10,
  num_mappings = 10,
  filter_variable = NULL,
  filter_concept = NULL,
  text_wrapping_char = 80,
  output_value = "prop_concept"
)
}
\arguments{
\item{process_output}{the output from \code{csd_process}}

\item{output_function}{the name of the output function that should be executed, provided in the message output
to the console after \code{csd_process} has been executed}

\item{concept_set}{the same concept set used in the \code{csd_process} function; only required if \code{vocab_tbl} is not NULL}

\item{vocab_tbl}{OPTIONAL: the location of an external vocabulary table containing concept names for
the provided codes. if not NULL, concept names will be available in either a reference
table or in a hover tooltip}

\item{num_variables}{an integer to represent the top number of variables to include for the exploratory analyses;
will pick based on the most commonly appearing variables;}

\item{num_mappings}{an integer to represent the top number of mappings for a given variable in the exploratory analyses;
will pick based on the highest count of the most commonly appearing variables;}

\item{filter_variable}{for both \verb{single- and multi- site anomaly tests without time measurements} and
\verb{single- and multi- site exploratory tests with time measurements}, the variables
to focus on}

\item{filter_concept}{for \code{ss_anom_la}, \code{ms_exp_la}, and \code{ms_anom_la}, the specific code that should
be the focus of the analysis}

\item{text_wrapping_char}{an integer to limit the length of text on an axis before wrapping is enforced;
used in \code{ms_anom_cs}}

\item{output_value}{the numerical column in \code{process_output} that should be used in the visualization
relevant for \code{ss_exp_la}, \code{ms_anom_cs}, and \code{ms_exp_la}}
}
\value{
a graph to visualize the results from \code{csd_process} based on the parameters provided; see documentation
for individual subfunctions for details on specific output
}
\description{
Using the tabular output generated by \code{csd_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_ssdqa()}
}
