% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_output.R
\name{csd_output}
\alias{csd_output}
\title{Concept Set Distribution -- Output Generation}
\usage{
csd_output(
  process_output,
  concept_set = NULL,
  vocab_tbl = NULL,
  num_variables = 10,
  num_mappings = 10,
  filter_variable = NULL,
  filter_concept = NULL,
  text_wrapping_char = 80,
  output_value = "prop_concept",
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{\emph{tabular input} || \strong{required}

The tabular output produced by \code{csd_process}}

\item{concept_set}{\emph{tabular input} || \strong{optional}

The concept set originally used in the \code{csd_process} function. Recommended
if no vocab_tbl is provided but the concept names are available in the concept set.}

\item{vocab_tbl}{\emph{tabular input} || \strong{optional}

A vocabulary table containing concept names for the provided codes
(ex: the OMOP concept table)}

\item{num_variables}{\emph{integer} || defaults to \code{30}

An integer indicating the top N of variables to include for the exploratory analyses.
The function will choose the most commonly occurring N variables to include in the plot.}

\item{num_mappings}{\emph{integer} || defaults to \code{30}

An integer indicating the top N of concepts to include for the exploratory analyses.
The function will choose the most commonly occurring N concepts per variable
to include in the plot.}

\item{filter_variable}{\emph{string or vector} || defaults to \code{NULL}

The specific variable(s) to display in the output. This parameter is required
for the following check types:
\itemize{
\item \verb{Single Site, Anomaly Detection, Cross-Sectional}
\item \verb{Multi Site, Anomaly Detection, Cross-Sectional}
\item \verb{Single Site, Exploratory, Longitudinal}
\item \verb{Single Site, Anomaly Detection, Longitudinal}
\item \verb{Multi Site, Exploratory, Longitudinal}
}}

\item{filter_concept}{\emph{numeric/string or vector} || defaults to \code{NULL}

The specific code(s) to display in the output. This parameter is required
for the following check types:
\itemize{
\item \verb{Single Site, Anomaly Detection, Longitudinal}
\item \verb{Multi Site, Exploratory, Longitudinal}
\item \verb{Multi Site, Anomaly Detection, Longitudinal}
}}

\item{text_wrapping_char}{\emph{integer} || defaults to \code{80}

An integer indicating the length limit for text on an axis before
wrapping is enforced. This is only used for the
\verb{Multi Site, Anomaly Detection, Cross-Sectional} check type}

\item{output_value}{\emph{string} || defaults to \code{prop_concept}

The name of the numerical column in \code{process_output} that should be
used in the output. This parameter is required for the following check
types:
\itemize{
\item \verb{Multi-Site, Anomaly Detection, Cross-Sectional}
\item \verb{Single Site, Exploratory, Longitudinal}
\item \verb{Multi-Site, Exploratory, Longitudinal}
}}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information.}
}
\value{
This function will produce a graph to visualize the results
from \code{csd_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using the tabular output generated by \code{csd_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'conceptsetdistribution'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'csd_process_test',
                      working_directory = getwd(),
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = system.file('extdata',
                                        package = 'conceptsetdistribution'),
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000),
                #RSQLite does not store date objects,
                #hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Prepare input tables
csd_domain_tbl <- dplyr::tibble(domain = 'condition_occurrence',
                                concept_field = 'condition_concept_id',
                                date_field = 'condition_start_date',
                                vocabulary_field = NA)

csd_concept_tbl <- read_codeset('dx_hypertension') \%>\%
  dplyr::mutate(domain = 'condition_occurrence',
                variable = 'hypertension')

#' Execute `csd_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
csd_process_example <- csd_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   domain_tbl = csd_domain_tbl,
                                   concept_set = csd_concept_tbl)

csd_process_example

#' Execute `csd_output` function
csd_output_example <- csd_output(process_output = csd_process_example,
                                 concept_set = csd_concept_tbl,
                                 vocab_tbl = NULL)

csd_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(csd_output_example[[1]])
}
