% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_csd_full.R, R/cohort_csd_full_pcnt.R
\name{csd_process}
\alias{csd_process}
\title{CSD Process function}
\usage{
csd_process(
  cohort,
  domain_tbl = read_codeset("scv_domains", "cccc"),
  concept_set = read_codeset("csd_codesets", "iccccc"),
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  num_concept_combined = FALSE,
  num_concept_1 = 30,
  num_concept_2 = 30,
  p_value = 0.9,
  age_groups = FALSE,
  time = TRUE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year"
)

csd_process(
  cohort,
  domain_tbl = read_codeset("scv_domains", "cccc"),
  concept_set = read_codeset("csd_codesets", "iccccc"),
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  num_concept_combined = FALSE,
  num_concept_1 = 30,
  num_concept_2 = 30,
  p_value = 0.9,
  age_groups = FALSE,
  time = TRUE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year"
)
}
\arguments{
\item{cohort}{cohort for SSDQA testing; required fields:
\code{site} | \code{person_id} | \code{start_date} | \code{end_date} where start and end date}

\item{domain_tbl}{tbl that is similar to the SCV check;
four columns: \code{domain} | \code{source_col} | \code{concept_col} | \code{date_col};
the required columns for the csd check are only \code{domain_tbl}, \code{concept_col}, \code{date_col}}

\item{concept_set}{concept set CSV file with the following columns:
\code{concept_id} | \code{concept_code} | \code{concept_name} | \code{vocabulary_id} | \code{category} | \code{variable} | \code{domain}
The variable field is required to categorize each concept set into a particular variable
The domain is required so that the function knows which table to join to in order to derive counts}

\item{multi_or_single_site}{direction to determine what kind of check to run
string that is either \code{multi} or \code{single}}

\item{anomaly_or_exploratory}{direction to determine what kind of check to run; a string
that is either \code{anomaly} or \code{exploratory}}

\item{num_concept_combined}{when \code{mult_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument is an integer and will ensure that \code{concept1} and \code{concept2} meet
some minimal threshold for including in the jaccard index; if \code{TRUE}, then
\emph{both} conditions for \code{num_concept_1} and \code{num_concept_2} should be met;
if \code{FALSE} then just one condition needs to be met.}

\item{num_concept_1}{when \code{mult_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument is an integer and requires a minimum number of times that
the \emph{first} concept appears in the dataset}

\item{num_concept_2}{when \code{mult_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument is an integer and requires a minimum number of times that
the \emph{second} concept appears in the dataset}

\item{p_value}{the p value to be used as a threshold in the multi-site anomaly detection analysis}

\item{age_groups}{N/A for this check?}

\item{time}{logical to determine whether to output the check across time}

\item{time_span}{when \code{time = TRUE}, a vector of two dates for the observation period of the study}

\item{time_period}{when time = TRUE, this argument defines the distance between dates within the specified time period. defaults
to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}
}
\value{
for \verb{single- and multi-} and \code{exploratory} analyses, the output is:
\code{site} | \code{variable} | \code{ct_denom} | \code{concept_id} | \code{ct_concept} | \code{prop_concept}
for \code{single} and \code{anomaly} analyses, the output is:
\code{site} | \code{concept1}| \code{concept2}| \code{cocount}| \code{concept1_ct} | \code{concept2_ct} | \code{concept_count_union} |\code{jaccard_index} |
\code{concept1_prop} | \code{concept2_prop} | \code{variable} |
where \code{concept_count_union} is how often a pair of codes appear in the full dataset
and \code{concept1_ct} and \code{concept2_ct} are how often each appear in the dataset
for any that are \verb{across time}, the output is:
\code{site} | \code{time_start} | \code{time_increment} | \code{variable} | \code{ct_denom} | \code{concept_id} | \code{ct_concept} | \code{prop_concept}

for \verb{single- and multi-} and \code{exploratory} analyses, the output is:
\code{site} | \code{variable} | \code{ct_denom} | \code{concept_id} | \code{ct_concept} | \code{prop_concept}
for \code{single} and \code{anomaly} analyses, the output is:
\code{site} | \code{concept1}| \code{concept2}| \code{cocount}| \code{concept1_ct} | \code{concept2_ct} | \code{concept_count_union} |\code{jaccard_index} |
\code{concept1_prop} | \code{concept2_prop} | \code{variable} |
where \code{concept_count_union} is how often a pair of codes appear in the full dataset
and \code{concept1_ct} and \code{concept2_ct} are how often each appear in the dataset
for any that are \verb{across time}, the output is:
\code{site} | \code{time_start} | \code{time_increment} | \code{variable} | \code{ct_denom} | \code{concept_id} | \code{ct_concept} | \code{prop_concept}
}
\description{
add csv output of main table and csv output of param summary
(that function needs updating to be more generalizable)

add csv output of main table and csv output of param summary
(that function needs updating to be more generalizable)
}
