% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd_process.R
\name{csd_process}
\alias{csd_process}
\title{Concept Set Distribution}
\usage{
csd_process(
  cohort,
  domain_tbl = conceptsetdistribution::csd_domain_file,
  concept_set = conceptsetdistribution::csd_concept_set,
  omop_or_pcornet,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  num_concept_combined = FALSE,
  num_concept_1 = 30,
  num_concept_2 = 30,
  p_value = 0.9,
  age_groups = NULL,
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year"
)
}
\arguments{
\item{cohort}{\emph{tabular input} || \strong{required}

The cohort to be used for data quality testing. This table should contain,
at minimum:
\itemize{
\item \code{site} | \emph{character} | the name(s) of institutions included in your cohort
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character} | the patient identifier
\item \code{start_date} | \emph{date} | the start of the cohort period
\item \code{end_date} | \emph{date} | the end of the cohort period
}

Note that the start and end dates included in this table will be used to
limit the search window for the analyses in this module.}

\item{domain_tbl}{\emph{tabular input} || \strong{required}

A table or CSV file defining the domains listed in the annotated concept set.
This input should contain four columns:
\itemize{
\item \code{domain} | \emph{character} | a string identifying the CDM table, as listed in the annotated concept set, where the concept of interest can be identified
\item \code{concept_field} | \emph{character}| the string name of the field in the domain table where the concepts are located
\item \code{date_field} | \emph{character} |  the name of the field in the domain table with the date that should be used for temporal filtering
\item \code{vocabulary_field} | \emph{character} | for PCORnet applications, the name of the field in the domain table with a vocabulary identifier to differentiate concepts from one another (ex: dx_type); can be set to NA for OMOP applications
}

To see an example of what this input should look like, see \code{?conceptsetdistribution::csd_domain_file}}

\item{concept_set}{\emph{tabular input} || \strong{required}

A table or CSV containing the concepts to be investigated, plus some additional
metadata. This input should contain one of following:
\itemize{
\item \code{concept_id} | \emph{integer} | the concept_id of interest (required for OMOP)
\item \code{concept_code} | \emph{character} | the code of interest (required for PCORnet)
}

And both of:
\itemize{
\item \code{variable} | \emph{character} | a string label grouping one concept code into a larger variable definition
\item \code{domain} | \emph{character} | the name of the CDM table where the concept can be found
}

For certain PCORnet applications, it should also contain
\itemize{
\item \code{vocabulary_id} | \emph{character} | the vocabulary of the code, which should match what is listed in the domain table's \code{vocabulary_field}
}

To see an example of what this input should look like, see \code{?conceptsetdistribution::csd_concept_set}}

\item{omop_or_pcornet}{\emph{string} || \strong{required}

A string, either \code{omop} or \code{pcornet}, indicating the CDM format of the data
\itemize{
\item \code{omop}: run the \code{\link[=csd_process_omop]{csd_process_omop()}} function against an OMOP CDM instance
\item \code{pcornet}: run the \code{\link[=csd_process_pcornet]{csd_process_pcornet()}} function against a PCORnet CDM instance
}}

\item{multi_or_single_site}{\emph{string} || defaults to \code{single}

A string, either \code{single} or \code{multi}, indicating whether a single-site or
multi-site analysis should be executed}

\item{anomaly_or_exploratory}{\emph{string} || defaults to \code{exploratory}

A string, either \code{anomaly} or \code{exploratory}, indicating what type of results
should be produced.

Exploratory analyses give a high level summary of the data to examine the
fact representation within the cohort. Anomaly detection analyses are
specialized to identify outliers within the cohort.}

\item{num_concept_combined}{\emph{boolean} || defaults to \code{FALSE}

When \code{multi_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument is a boolean that will ensure that \code{concept1} and \code{concept2}
meet some minimal threshold for inclusion in the Jaccard index computation.

if \code{TRUE}, then \emph{both} conditions for \code{num_concept_1} and \code{num_concept_2}
should be met; if \code{FALSE} then just one condition needs to be met.}

\item{num_concept_1}{\emph{integer} || defaults to \code{30}

When \code{multi_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument indicates the minimum number of times that the \emph{first} concept
appears in the dataset during the Jaccard index computation}

\item{num_concept_2}{\emph{integer} || defaults to \code{30}

When \code{multi_or_single_site} = \code{single} and \code{anomaly_or_exploratory} = \code{anomaly},
this argument indicates the minimum number of times that the \emph{second} concept
appears in the dataset during the Jaccard index computation}

\item{p_value}{\emph{numeric} || defaults to \code{0.9}

The p value to be used as a threshold in the Multi-Site,
Anomaly Detection, Cross-Sectional analysis}

\item{age_groups}{\emph{tabular input} || defaults to \code{NULL}

If you would like to stratify the results by age group, create a table or
CSV file with the following columns and use it as input to this parameter:
\itemize{
\item \code{min_age} | \emph{integer} | the minimum age for the group (i.e. 10)
\item \code{max_age} | \emph{integer} | the maximum age for the group (i.e. 20)
\item \code{group} | \emph{character} | a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave as \code{NULL}}

\item{time}{\emph{boolean} || defaults to \code{FALSE}

A boolean to indicate whether to execute a longitudinal analysis}

\item{time_span}{\emph{vector - length 2} || defaults to \code{c('2012-01-01', '2020-01-01')}

A vector indicating the lower and upper bounds of the time series for longitudinal analyses}

\item{time_period}{\emph{string} || defaults to \code{year}

A string indicating the distance between dates within the specified time_span.
Defaults to \code{year}, but other time periods such as \code{month} or \code{week} are
also acceptable}
}
\value{
This function will return a dataframe summarizing the
distribution of code usage for each user defined variable. For a
more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
This is an information representation module that will compute the distribution of concept usage in a
user-provided annotated concept set. The user will define the domains of interest (\code{domain_tbl}) and
provide the concept set to be evaluated (\code{concept_set}). Sample versions of these inputs are
included as data in the package and are accessible with \verb{conceptsetdistribution::}.
Results can optionally be stratified by site, age group, and/or time.
This function is compatible with both the OMOP and the PCORnet CDMs based on the
user's selection.
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'conceptsetdistribution'))

getwd()

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'csd_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000),
                #RSQLite does not store date objects,
                #hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Prepare input tables
csd_domain_tbl <- dplyr::tibble(domain = 'condition_occurrence',
                                concept_field = 'condition_concept_id',
                                date_field = 'condition_start_date',
                                vocabulary_field = NA)

csd_concept_tbl <- read_codeset('dx_hypertension') \%>\%
  dplyr::mutate(domain = 'condition_occurrence',
                variable = 'hypertension')

#' Execute `csd_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
csd_process_example <- csd_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   domain_tbl = csd_domain_tbl,
                                   concept_set = csd_concept_tbl) \%>\%
  suppressMessages()

csd_process_example

#' Execute `csd_output` function
csd_output_example <- csd_output(process_output = csd_process_example,
                                 concept_set = csd_concept_tbl,
                                 vocab_tbl = NULL) \%>\%
  suppressMessages()

csd_output_example[[1]]
csd_output_example[[2]]

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(csd_output_example[[1]])
}
